# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CareConnect API
  description: >
    Clinic appointment and patient-triage API. Implements the contract
    described in docs/spec.md. This file is the single source of truth: on
    any conflict with another doc, this file wins. All endpoints require
    staff authentication (front-desk or doctor) except login — there is no
    patient-facing surface in v1.


    Conventions: all timestamps are UTC ISO-8601 date-times; `date`
    parameters and `ClockTime` values are interpreted in
    `ClinicSettings.clinicTimezone`. Status mapping — 400 malformed or
    schema-invalid request, 401 unauthenticated, 403 permission denied,
    404 not found, 409 state conflict, 422 policy/business-rule violation,
    429 rate limited, 500 unexpected server error (`INTERNAL_ERROR`, no
    details), 503 AI provider unavailable. Every error body is
    `Error` with a `code` from `ErrorCode`.
  version: 0.2.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /specialties:
    get:
      tags: [Specialties]
      summary: List specialties
      responses:
        "200":
          description: Specialties
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Specialty" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Specialties]
      summary: Create a specialty
      description: Front-desk only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpecialtyCreate" }
      responses:
        "201":
          description: Specialty created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Specialty" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /specialties/{specialtyId}:
    parameters:
      - name: specialtyId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Specialties]
      summary: Update a specialty
      description: Front-desk only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpecialtyUpdate" }
      responses:
        "200":
          description: Updated specialty
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Specialty" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------ Patients
  /patients:
    get:
      tags: [Patients]
      summary: Look up or list patients
      description: >
        With `phone`, returns every patient registered under that number
        (spec.md §2 — a phone may map to multiple patients, e.g. family
        members). Without it, returns a general paginated list. Doctors
        may read all patients (spec.md §1).
      parameters:
        - name: phone
          in: query
          description: Any common phone format; normalized to E.164 before matching.
          schema: { type: string }
        - name: name
          in: query
          description: Case-insensitive name prefix filter.
          schema: { type: string }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Matching patients
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PatientPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Patients]
      summary: Register a patient
      description: >
        Rejected with `PATIENT_ALREADY_EXISTS` if the (E.164 phone,
        normalized name) pair already exists.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatientCreate" }
      responses:
        "201":
          description: Patient registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /patients/{patientId}:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [Patients]
      summary: Get a patient
      responses:
        "200":
          description: The patient
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Patients]
      summary: Update patient demographics
      description: >
        Front-desk or a doctor (spec.md §1). A change that collides with
        an existing (phone, name) is rejected with
        `PATIENT_ALREADY_EXISTS`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatientUpdate" }
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /patients/{patientId}/medical-history:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [Patients]
      summary: List a patient's medical history entries
      responses:
        "200":
          description: Medical history (free-text, append-only, per spec.md §2)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MedicalHistoryEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Patients]
      summary: Add a medical history entry or an amendment
      description: >
        Append-only. To correct an entry, post `kind=amendment` with
        `amendsEntryId`; the original is never edited or deleted.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MedicalHistoryEntryCreate" }
      responses:
        "201":
          description: Entry added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MedicalHistoryEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- Slots
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PatientId:
      name: patientId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Malformed or schema-invalid request (code VALIDATION_ERROR)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or expired credentials, or deactivated user
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted (see spec.md §1)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: >
        State conflict — see the operation description for the specific
        `ErrorCode` (e.g. SLOT_ALREADY_BOOKED, PATIENT_ALREADY_BOOKED,
        PATIENT_ALREADY_EXISTS, DOCTOR_ALREADY_EXISTS, SPECIALTY_ALREADY_EXISTS,
        AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS, AVAILABILITY_OVERLAP, USER_ALREADY_EXISTS, APPOINTMENT_NOT_COMPLETED,
        VISIT_NOTE_NOT_WRITABLE, VISIT_NOTE_LOCKED)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ErrorCode:
      type: string
      description: >
        The catalog of error codes. Each code has exactly one meaning and is
        used consistently everywhere it is raised.
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - INVALID_CREDENTIALS
        - FORBIDDEN
        - NOT_FOUND
        - RATE_LIMITED
        - SLOT_ALREADY_BOOKED
        - PATIENT_ALREADY_BOOKED
        - INVALID_SLOT
        - EMERGENCY_JUSTIFICATION_REQUIRED
        - EMERGENCY_NOT_AUTHORIZED
        - CANCELLATION_WINDOW_CLOSED
        - INVALID_TRANSITION
        - FOLLOW_UP_WINDOW_EXCEEDED
        - APPOINTMENT_NOT_COMPLETED
        - PATIENT_ALREADY_EXISTS
        - USER_ALREADY_EXISTS
        - DOCTOR_ALREADY_EXISTS
        - SPECIALTY_ALREADY_EXISTS
        - AVAILABILITY_OVERLAP
        - AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS
        - VISIT_NOTE_NOT_WRITABLE
        - VISIT_NOTE_LOCKED
        - NO_NOTES_TO_DRAFT
        - AI_SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
    Error:
      type: object
      required: [code, message]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
...
    Specialty:
      type: object
      required: [id, name, defaultSlotLengthMinutes]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        defaultSlotLengthMinutes: { type: integer, minimum: 5 }
    SpecialtyCreate:
      type: object
      required: [name, defaultSlotLengthMinutes]
      properties:
        name: { type: string }
        defaultSlotLengthMinutes: { type: integer, minimum: 5 }
    SpecialtyUpdate:
      type: object
      properties:
        name: { type: string }
        defaultSlotLengthMinutes: { type: integer, minimum: 5 }
...
    Patient:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone:
          type: string
          description: >
            Stored as E.164. Identity is the (phone, normalized name)
            pair, which is unique — one phone number may have multiple
            registered patients (spec.md §2).
        dob: { type: string, format: date, nullable: true }
        email: { type: string, format: email, nullable: true }
        createdAt: { type: string, format: date-time }
    PatientCreate:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        dob: { type: string, format: date, nullable: true }
        email: { type: string, format: email, nullable: true }
    PatientUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        dob: { type: string, format: date, nullable: true }
        email: { type: string, format: email, nullable: true }
    PatientPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Patient" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    MedicalHistoryKind:
      type: string
      enum: [entry, amendment]
    MedicalHistoryEntry:
      type: object
      required: [id, patientId, kind, description, recordedAt, recordedBy]
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/MedicalHistoryKind" }
        amendsEntryId: { type: string, format: uuid, nullable: true }
        description:
          type: string
          description: Free text (spec.md §2) — no structured fields.
        recordedAt: { type: string, format: date-time }
        recordedBy: { type: string, format: uuid }
    MedicalHistoryEntryCreate:
      type: object
      required: [description]
      properties:
        kind:
          allOf: [{ $ref: "#/components/schemas/MedicalHistoryKind" }]
          default: entry
        amendsEntryId:
          type: string
          format: uuid
          nullable: true
          description: Required when kind is `amendment`.
        description: { type: string, minLength: 1 }
...